remove bucketId from dto metedata, remove namedMetadata class
