+++ src/integration-test/java/org/ow2/proactive/catalog/rest/controller/CatalogObjectRevisionControllerIntegrationTest.java
@@ -98,9 +98,7 @@ public class CatalogObjectRevisionControllerIntegrationTest extends AbstractCata
                                                 .extract()
                                                 .path("");
 
-        bucket = new BucketMetadata(((Integer) result.get("id")).longValue(),
-                                    (String) result.get("name"),
-                                    (String) result.get("owner"));
+        bucket = new BucketMetadata((String) result.get("name"), (String) result.get("owner"));
 
         // Add an object of kind "workflow" into first bucket
         catalogObjectRevisionAlone = given().pathParam("bucketName", bucket.getName())
